Move mapping file into fty-common-nut
