--- src/Makemodule-local.am
@@ -1,20 +1,0 @@
-EXTRA_DIST += \
-	src/mapping.conf \
-	src/alert_device_alert.h
-
-AM_CPPFLAGS += \
-	-I$(srcdir)/src
-
-CLEANFILES += \
-	stderr.txt
-
-$(abs_builddir)/src/mapping.conf: $(abs_srcdir)/src/mapping.conf
-	@if [ "$@" != "$<" ]; then cp -pf "$<" "$@" ; fi
-
-check-local: $(abs_builddir)/src/mapping.conf
-
-clean-local: clean-local-check
-.PHONY: clean-local-check
-clean-local-check:
-	rm -rf .testdir || true
-	@if [ "$(abs_builddir)" != "$(abs_srcdir)" ]; then rm -f $(abs_builddir)/src/mapping.conf ; fi
