+++ src/selftest-ro/mapping.conf
@@ -1,0 +1,31 @@
+{
+    "physicsMapping" : {
+        "battery.charge"        :   "charge.battery",
+        "battery.runtime"       :   "runtime.battery",
+        "battery.voltage"       :   "voltage.battery",
+
+        "outlet.#.current"      :   "current.outlet.#",
+        "outlet.#.voltage"      :   "voltage.outlet.#",
+        "outlet.#.realpower"    :   "realpower.outlet.#",
+        "outlet.#.power"        :   "power.outlet.#",
+        "outlet.group.#.current"   :   "current.outlet.group.#",
+        "outlet.group.#.voltage"   :   "voltage.outlet.group.#",
+        "outlet.group.#.realpower" :   "realpower.outlet.group.#",
+        "outlet.group.#.power"     :   "power.outlet.group.#",
+        "outlet.group.#.load"      :   "load.outlet.group.#"
+    },
+    "inventoryMapping" : {
+        "device.model"          :       "model",
+        "ups.model"             :       "model",
+        "device.mfr"            :       "manufacturer",
+        "ups.mfr"               :       "manufacturer",
+        "ups.status"            :       "status.ups",
+        "ups.alarm"             :       "ups.alarm",
+        "ups.serial"            :       "serial_no",
+
+        "outlet.count"          :       "outlet.count",
+        "outlet.switchable"     :       "outlet.switchable",
+        "outlet.#.id"           :       "outlet.#.id",
+        "outlet.#.status"        :      "status.outlet.#"
+    }
+}
